Modified to also run on chrome
require InteropChrome.
Attempts to guess the the correct unified plan response when firefox is offering
unified plan to chrome. This gets interesting, because if firefox offers a second
video stream that is send only, chrome implicitly accepts it without modifying its
plan b sdp. So we have to use the offer and chromes last response to rebuild an appropriate
m=line. Not sure if this is guarenteed to work in all cases, but it works for us.